Move ShardingSphereDataCollector from pipeline to infra.
