+++ kernel/data-pipeline/core/src/test/resources/META-INF/services/org.apache.shardingsphere.infra.metadata.data.collector.ShardingSphereDataCollector
